Updated proxy scripts

+++ helpers/squid-helper.sh
@@ -7,12 +7,13 @@
 #############################################################################
 # Variables for our script:
 SQUID_CERT_DIR=/etc/squid/cert
-SQUID_CERT=${SQUID_CERT_DIR}/squid_proxyCA.pem
+SQUID_CERT=${SQUID_CERT_DIR}/ca.pem
+DHPARAM=${SQUID_CERT_DIR}/dhparam.pem
 CA_DIR=/usr/local/share/ca-certificates
 CA_CERT=${CA_DIR}/squid_proxyCA.crt
 
 # Return with error code 0 if this process is already done:
-test -f ${SQUID_CERT} && test -f ${CA_CERT} && exit 0
+test -f ${SQUID_CERT/pem/key} && test -f ${SQUID_CERT} && test -f ${SQUID_CERT/pem/der} && test -f ${DHPARAM} && test -f ${CA_CERT} && exit 0 
 
 # We need to be root in order to execute everything after this:
 if [[ "${UID}" -ne 0 ]]; then
@@ -21,8 +22,15 @@ if [[ "${UID}" -ne 0 ]]; then
 fi
 
 # Create the openssl configuration file if it doesn't already exist:
-rm -rf ${SQUID_CERT_DIR}
 mkdir -p ${SQUID_CERT_DIR}
+test -f ${SQUID_CERT} && rm ${SQUID_CERT}
+
+# Generate a 2048-bit self-signed certificate key:
+openssl genrsa -out ${SQUID_CERT/pem/key} 2048
+if [[ $? -ne 0 ]]; then echo "ERROR: Failed to generate ${SQUID_CERT/pem/key}"; exit 1; fi
+
+# Generate configuration file to use with openssl.  If we couldn't get information about
+# our IP, list the requested information is "Unspecified" and the country as "US":
 CONFIG=${SQUID_CERT_DIR}/openssl.config
 if [[ ! -f ${CONFIG} ]]; then
 	# Try to get information from ipinfo.io about our IP address:
@@ -31,10 +39,8 @@ if [[ ! -f ${CONFIG} ]]; then
 	STATE=$(grep "region" ${IPINFO} | cut -d\" -f 4)
 	CITY=$(grep "city" ${IPINFO} | cut -d\" -f 4)
 	COUNTRY=$(grep "country" ${IPINFO} | cut -d\" -f 4)
-	rm ${IPINFO}
+	test -f ${IPINFO} && rm ${IPINFO}
 
-	# Generate configuration file to use with openssl.  If we couldn't get information about
-	# our IP, list the requested information is "Unspecified" and the country as "US":
 	(
 	echo "[req]"
 	echo "default_bit         = 4096"
@@ -45,28 +51,31 @@ if [[ ! -f ${CONFIG} ]]; then
 	echo "countryName         = ${COUNTRY:-"US"}"
 	echo "stateOrProvinceName = ${STATE:-"Unspecified"}"
 	echo "localityName        = ${CITY:-"Unspecified"}"
-	echo "organizationName    = BPiWRT"
+	echo "organizationName    = Self-signed CA"
 	) > ${CONFIG}
 fi
 
-# Generate self-signed CA certificate/key, both in the same file:
-openssl req -new -newkey rsa:4096 -sha256 -days 3650 -nodes -x509 -keyout ${SQUID_CERT} -out ${SQUID_CERT} -config ${CONFIG}
-if [[ $? -ne 0 ]]; then
-	echo "ERROR: Failed to generate ${SQUID_CERT}"
-	exit 1
-fi
+# Generate the certificate PEM file:
+openssl req -x509 -new -nodes -sha256 -days 3650 -key ${SQUID_CERT/pem/key} -out ${SQUID_CERT}  -config ${CONFIG}
+if [[ $? -ne 0 ]]; then echo "ERROR: Failed to generate ${SQUID_CERT}"; exit 2; fi
+
+# Generate the DER file used from the PEM file: 
+openssl x509 -in ${SQUID_CERT} -outform DER -out ${SQUID_CERT/pem/der}
+if [[ $? -ne 0 ]]; then echo "ERROR: Failed to generate ${SQUID_CERT/pem/der}"; exit 3; fi
+
+# Download a randomly selected 2048-bit dhparam file.  If that fails, generate one! 
+# This takes long time, though.  If it still fails, abort with error message! 
+wget -q -O ${DHPARAM} https://2ton.com.au/getprimes/random/dhparam/2048
+[[ $? -ne 0 ]] && openssl dhparam -outform PEM -out ${DHPARAM} 2048
+if [[ $? -ne 0 ]]; then echo "ERROR: Failed to generate ${DHPARAM}"; exit 4; fi
+
+# Change ownership and visibility of generated certificates:
 chown -R proxy:proxy ${SQUID_CERT_DIR}
 chmod 0400 ${SQUID_CERT}*
 
 # Add squid_proxyCA cert to system so it's trusted by default:
-rm -rf ${CA_DIR}
 mkdir -p ${CA_DIR}
+test -f ${CA_CERT} && rm ${CA_CERT}
 openssl x509 -inform PEM -in ${SQUID_CERT} -out ${CA_CERT}
-if [[ $? -ne 0 ]]; then
-	echo "ERROR: Failed to generate ${CA_CERT}"
-	exit 2
-fi
+if [[ $? -ne 0 ]]; then echo "ERROR: Failed to generate ${CA_CERT}"; exit 5; fi
 update-ca-certificates
-
-# Call the "privoxy-blocklist.sh" script to load everything we need:
-/etc/privoxy/privoxy-blocklist.sh
